requirements.txt
python-dotenv==1.0.0
watchdog==6.0.0
Flask==3.1.1